financial_engine/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.25
yfinance>=0.2
pandas>=2.1
aiofiles>=23.2
python-multipart>=0.0.7
pytz>=2024
apscheduler>=3.10
google-cloud-storage>=2.0.0
google-auth>=2.29.0
httpx>=0.27
